Changed onMessage
When returning from sendMessage, Firefox seems to convert anything that was thrown into a standard Error, with the message property of the Error getting the message property of what was thrown, if it was an object with a "truthy" message property, or "An unexpected error occurred" otherwise.
I don't think Firefox returns anything else besides the message (i.e. not the type of error, filename, linenumber, etc.).
To allow the same functionality, the message property is returned in an error-like object, to be converted to an Error on arrival.