Rename some of the things
An attempt to apply a consistent naming scheme to some major functions and data, to hopefully clarify the code.  Particular changes are:

The terms "wrap" and "wrappers" seemed to be used to apply to two kinds of things.  I have divided these into:
- proxies: things that actually polyfill parts of the Firefox API
- wrappers: things that wrap parts of the Chrome API, as an intermediate step to polyfilling parts of the Firefox API

Functions that return a new proxy or wrapper for some X are generally called "wrapX".  I think "new_XProxy" / "new_XWrapper" may be clearer.

A couple of functions are declared with the intention of being bound before use.  I think naming these "unbound_..." makes this clearer.

Also, changed to pass chrome as an argument to the main function, as per #139.