Fix no-op check to work with Safari and support of old browsers.
Verify the object that might be at `globalThis.browser` is not already
implementing the basic Web Extension APIs. This is symmetric with the
check to verify the poly fill is only included in extension contexts.

Don't use optional-chaining for these checks, since this needs to
continuing working in older browsers.