fix: Remove webpack and webpack-cli dependencies
Changes made in #597 accidentally resulted in `webpack` and
`webpack-cli` being added as dependencies in the manifest of the
published package. This problem can be seen in v0.11.0. For more
information, see #604.

`webpack` and `webpack-cli` are used for bundler smoke tests that are
run in CI. Rather than being installed during CI, these dependencies
are now included in `devDependencies`. A new npm script has been added
for running this script as well, to ensure that these dependencies are
on the PATH when this script is run.

The CircleCI config and the contributor docs have been updated to use
the new npm script. You can test it by running
`npm run test-module-bundlers-smoketest`.

Fixes #604